requirements.txt
django==5.0.12
mssql-django==1.5
djangorestframework==3.15.2
rest_framework_simplejwt==0.0.2
Pillow==11.1.0